Change int(x[1]) to float(x[1]) in sorted_entries
Line 719, handles decimals in database time (seconds)

+++ gameHandler.py
@@ -716,7 +716,7 @@ def dummyReturn():
     return
 
 def display_leaderboard(screen, data):
-    sorted_entries = sorted(data, key=lambda x: int(x[1]))
+    sorted_entries = sorted(data, key=lambda x: float(x[1]))
     
     no_duplicates, logged = [], []
     for entry in sorted_entries:
